Add configuration file.
This is based on Jacek Izykowski's work:
https://github.com/Juniper/contrail-windows-docker-driver/pull/21

Initial patchset is a merge of his work to master. Following patchsets are fixes and
improvements.

Changes:
* Remove loading keystone variables from environment.
* Introduce an ini-format config file.
* Command line params take precedence over config params.
* Move flags to configuration/ package.
* Refactors.
* Introduce dryRun flag that parses config but doesn't do anything.
* Bake sample test config file into test binary using go generate. This
is done so that we don't have to complicate integration test setup on
CI-side and needlessly couple it. It will be easier to add test cases
this way as well.

Co-authored-by: Jacek Izykowski <jacek.izykowski@codilime.com>
Co-authored-by: Michal Kostrzewa <michal.kostrzewa@codilime.com>

Change-Id: I92f0df4f84f519836a0af183cc4580f72189feb0
Closes-Bug: #1787565