Build MSI installer in Invoke-Build.ps1.
It's very hard to introduce Invoke-Build.ps1 to CI without also
providing a way to build the MSI. The best way to fix it is to introduce
a way to build the MSI directly in the docker driver repository.

We create wix_new.json file for now. We change the path specified inside
it. If we modify wix.json, CI fails. We can throw out wix.json and
rename wix_new.json to wix.json later - after we start using
./Invoke-Build.ps1 in CI.

Change-Id: I3dc380ede3ad52890e4e9763569fbcd8833915fd
Partial-Bug: #1777423