Refactor out CreateEndpoint method.
CreateEndpoint touches many different parts of the codebase, so the
refactors are surprisingly extensive.

* Create a facade for Controller to try and shrink the Controller port
interface. It's still WIP (needs to be done gradually when refactoring
other handlers), but I belive it's a step in the right direction.
* Refactor CreateEndpoint by simplifying the way it calls its
collaborators. Move it to driver_core package.
* Fixed fake controller implementation by using
contrail-go-api Interceptors to mimic the actual API server behaviour of
assigning Default GW, MAC and IPs. This accidentally fixed two
controlelr_test.go tests. Oops.

Change-Id: I8a4f9dbdd390f7167409689cd2ae6c204b0a3edb
Partial-Bug: #1778671