Bump hcsshim and go-winio version.
They are being bumped simultaneusly due to dependency between them
(hcsshim imports go-winio).

Bump of go-winio is a possible fix to #1787946. Reopen this issue if the
bug happens again.

Bump of hcsshim, because it's a core dependency and it wasn't updated in
a couple months. Maybe some bugs were fixed.

Change-Id: I7cea4c004a5064f03b2b9fff8d99032babcf3137
Closes-Bug: #1787950
Closes-Bug: #1787946